@@ -1,10 +1,11 @@
 {
   "name": "react-tetris",
   "version": "1.0.0",
-  "description": "使用React、Redux、Immutable编写「俄罗斯方块」。Use Tetact, Redux, Immutable to coding \"Tetris\".",
+  "description": "使用 React、Redux、Immutable 编写「俄罗斯方块」。Use Tetact, Redux, Immutable to coding \"Tetris\".",
   "scripts": {
     "start": "webpack-dev-server --progress",
-    "build": "rm -rf ./build/* && webpack --config ./webpack.production.config.js --progress && ls ./build"
+    "build": "rm -rf ./build/* && webpack --config ./webpack.production.config.js --progress && ls ./build",
+    "dev": "node build/dev-server.js"
   },
   "repository": {
     "type": "git",
@@ -24,43 +25,65 @@
   },
   "homepage": "https://github.com/chvin/react-tetris#readme",
   "devDependencies": {
-    "autoprefixer": "^6.7.2",
-    "babel-core": "^6.13.2",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-react-transform": "^2.0.2",
-    "babel-preset-es2015": "^6.13.2",
-    "babel-preset-react": "^6.16.0",
-    "copy-webpack-plugin": "^3.0.1",
-    "css-loader": "^0.23.1",
-    "eslint": "^3.3.1",
-    "eslint-config-airbnb": "^10.0.1",
-    "eslint-loader": "^1.6.1",
-    "eslint-plugin-import": "^1.13.0",
-    "eslint-plugin-jsx-a11y": "^2.1.0",
-    "eslint-plugin-react": "^6.1.1",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^2.22.0",
+    "autoprefixer": "^6.7.7",
+    "babel-core": "^6.24.0",
+    "babel-loader": "^6.4.1",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "chalk": "^1.1.3",
+    "compression-webpack-plugin": "^0.4.0",
+    "connect-history-api-fallback": "^1.3.0",
+    "copy-webpack-plugin": "^4.0.1",
+    "css-loader": "^0.28.0",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-friendly-formatter": "^2.0.7",
+    "eslint-loader": "^1.7.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.10.3",
+    "eventsource-polyfill": "^0.9.6",
+    "express": "^4.15.2",
+    "extract-text-webpack-plugin": "^2.1.0",
+    "file-loader": "^0.11.1",
+    "friendly-errors-webpack-plugin": "^1.6.1",
+    "html-webpack-plugin": "^2.28.0",
+    "http-proxy-middleware": "^0.17.4",
     "json-loader": "^0.5.4",
-    "less": "^2.7.1",
-    "less-loader": "^2.2.3",
-    "open-browser-webpack-plugin": "0.0.3",
-    "postcss": "^5.2.12",
-    "postcss-loader": "^1.2.2",
+    "less": "^2.7.2",
+    "less-loader": "^4.0.3",
+    "open-browser-webpack-plugin": "0.0.5",
+    "opn": "^4.0.2",
+    "optimize-css-assets-webpack-plugin": "^1.3.1",
+    "ora": "^1.2.0",
+    "postcss": "^5.2.17",
+    "postcss-loader": "^1.3.3",
     "precss": "^1.4.0",
     "react-transform-hmr": "^1.0.4",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1"
+    "rimraf": "^2.6.1",
+    "semver": "^5.3.0",
+    "style-loader": "^0.16.1",
+    "url-loader": "^0.5.8",
+    "webpack": "^2.4.1",
+    "webpack-bundle-analyzer": "^2.4.0",
+    "webpack-dev-middleware": "^1.10.1",
+    "webpack-dev-server": "^2.4.2",
+    "webpack-hot-middleware": "^2.18.0",
+    "webpack-merge": "^4.1.0"
   },
   "dependencies": {
     "classnames": "^2.2.5",
     "immutable": "^3.8.1",
-    "react": "^15.3.0",
-    "react-dom": "^15.3.0",
-    "react-redux": "^4.4.5",
-    "redux": "^3.5.2",
-    "redux-immutable": "^3.0.8"
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-redux": "^5.0.4",
+    "redux": "^3.6.0",
+    "redux-immutable": "^4.0.0"
+  },
+  "engines": {
+    "node": ">= 6.10.0",
+    "npm": ">= 4.5.0"
   }
 }
