@@ -45,8 +45,8 @@ export default class Pause extends React.Component {
             bg: true,
             [style.pause]: true,
             [style.c]: this.state.showPause,
-          }
-        )}
+          })
+        }
       />
     );
   }
