@@ -6,13 +6,12 @@ import App from './containers/';
 import './unit/const';
 import './control';
 import { subscribeRecord } from './unit';
+import './resource/css/loader.css';
 
-subscribeRecord(store); // 将更新的状态记录到localStorage
+subscribeRecord(store); // 将更新的状态记录到 localStorage
 
 render(
   <Provider store={store}>
     <App />
-  </Provider>
-    , document.getElementById('root')
-);
-
+  </Provider>,
+  document.getElementById('root'));
