@@ -1,6 +1,9 @@
 import { createStore } from 'redux';
 import rootReducer from '../reducers';
 
-const store = createStore(rootReducer, window.devToolsExtension && window.devToolsExtension());
+/* eslint-disable no-underscore-dangle */
+const store = createStore(rootReducer,
+                          window.__REDUX_DEVTOOLS_EXTENSION__ &&
+                          window.__REDUX_DEVTOOLS_EXTENSION__());
 
 export default store;
